transfer_argument
refactoring: fix transfer_argument func

+++ main.h
@@ -4,7 +4,7 @@
 struct Node
 {   
     int   type   ;
-    int   data   ;
+    long  data   ;
     Node* left   ;
     Node* right  ;
     Node* pointer;
